requirements.txt
pyyaml
flask
pydantic
gunicorn